Add hardcoded projection for exportcube
I am  trying to add the possibility to add a projection to the data
cube.
I first wanted to test this with a fixed projection.
This would help in using non epsg:4326 projected datacubes, because we
can already include them, but the spatial coordinate reference system
gets lost in the datacube.